# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/albums/{albumId}": {
      "get": {
        "operationId": "AlbumsController_getAlbum_v1",
        "parameters": [
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The album."
          },
          "404": {
            "description": "No album has this ID."
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
    "/v1/albums/{albumId}/tracks": {
      "get": {
        "operationId": "AlbumsController_listTracks_v1",
        "parameters": [
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of the album's tracks with notes, most popular first."
          },
          "400": {
            "description": "Invalid `cursor` or `limit`."
          },
          "404": {
            "description": "No album has this ID."
          }
        },
        "tags": [
          "albums"
        ]
      }
    }
  },
  "info": {
    "title": "notefinder API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
...
}